\name{mlpredict-package}
\alias{mlpredict-package}
\alias{mlpredict}
\docType{package}
\title{
\packageTitle{mlpredict}
}
\description{
\packageDescription{mlpredict}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlpredict}
\packageIndices{mlpredict}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mlpredict}

Maintainer: \packageMaintainer{mlpredict}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
